✨Feat: Exhibition API 1차 완료
* :recycle:Refactor: 필드명 변경

* :recycle:Refactor: 오프라인 전시 설명 필드 추가

* :recycle:Refactor: 오류 메세지 수정

* :recycle:Refactor: 전시 좋아요 매핑테이블 필드 추가

* :card_file_box:Comment: JavaDocs 상세 정보 작성

* :sparkles:Feat: Exhibition API 완료

* :recycle:Refactor: 코드 포맷 적용

* :recycle:Refactor: 코드 리뷰 기반 수정사항 반영

* :recycle:Refactor: 코드 리뷰 기반 수정사항 반영

* :recycle:Refactor: 코드 포맷 적용

+++ src/main/java/com/likelion13/artium/domain/exhibition/service/ExhibitionServiceImpl.java
@@ -5,6 +5,8 @@ package com.likelion13.artium.domain.exhibition.service;
 
 import java.time.LocalDate;
 import java.time.ZoneId;
+import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -16,13 +18,18 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.likelion13.artium.domain.exhibition.dto.request.ExhibitionRequest;
+import com.likelion13.artium.domain.exhibition.dto.response.ExhibitionDetailResponse;
 import com.likelion13.artium.domain.exhibition.dto.response.ExhibitionResponse;
 import com.likelion13.artium.domain.exhibition.entity.Exhibition;
 import com.likelion13.artium.domain.exhibition.entity.ExhibitionStatus;
+import com.likelion13.artium.domain.exhibition.entity.SortBy;
 import com.likelion13.artium.domain.exhibition.exception.ExhibitionErrorCode;
 import com.likelion13.artium.domain.exhibition.mapper.ExhibitionMapper;
+import com.likelion13.artium.domain.exhibition.mapping.ExhibitionParticipant;
 import com.likelion13.artium.domain.exhibition.repository.ExhibitionRepository;
 import com.likelion13.artium.domain.user.entity.User;
+import com.likelion13.artium.domain.user.exception.UserErrorCode;
+import com.likelion13.artium.domain.user.repository.UserRepository;
 import com.likelion13.artium.domain.user.service.UserService;
 import com.likelion13.artium.global.exception.CustomException;
 import com.likelion13.artium.global.page.mapper.PageMapper;
@@ -39,6 +46,7 @@ import lombok.extern.slf4j.Slf4j;
 public class ExhibitionServiceImpl implements ExhibitionService {
 
   private final ExhibitionRepository exhibitionRepository;
+  private final UserRepository userRepository;
   private final UserService userService;
   private final S3Service s3Service;
   private final ExhibitionMapper exhibitionMapper;
@@ -46,41 +54,96 @@ public class ExhibitionServiceImpl implements ExhibitionService {
 
   @Override
   @Transactional
-  public String createExhibition(MultipartFile image, ExhibitionRequest request) {
+  public ExhibitionDetailResponse createExhibition(MultipartFile image, ExhibitionRequest request) {
 
     if (request.getStartDate().isAfter(request.getEndDate())) {
       throw new CustomException(ExhibitionErrorCode.INVALID_DATE_RANGE);
     }
 
     String imageUrl = null;
-    ExhibitionStatus status = ExhibitionStatus.UPCOMING;
-    LocalDate now = LocalDate.now(ZoneId.of("Asia/Seoul"));
+    ExhibitionStatus status = determineStatus(request.getStartDate(), request.getEndDate());
 
     if (image != null) {
       imageUrl = s3Service.uploadFile(PathName.EXHIBITION, image);
     }
 
-    if (request.getEndDate().isBefore(now) || request.getEndDate().isEqual(now)) {
-      status = ExhibitionStatus.ENDED;
-    } else if (request.getStartDate().isBefore(now) || request.getStartDate().isEqual(now)) {
-      status = ExhibitionStatus.ONGOING;
-    }
+    List<ExhibitionParticipant> participants = buildParticipants(request.getUserIdList());
+    User currentUser = userService.getCurrentUser();
 
-    Exhibition exhibition = exhibitionMapper.toExhibition(imageUrl, request, status);
+    Exhibition exhibition =
+        exhibitionMapper.toExhibition(imageUrl, request, status, currentUser, participants);
+
+    participants.forEach(p -> p.setExhibition(exhibition));
 
     try {
       exhibitionRepository.save(exhibition);
     } catch (Exception e) {
-      s3Service.deleteFile(s3Service.extractKeyNameFromUrl(imageUrl));
+      if (imageUrl != null) {
+        s3Service.deleteFile(s3Service.extractKeyNameFromUrl(imageUrl));
+      }
       throw new CustomException(ExhibitionErrorCode.EXHIBITION_API_ERROR);
     }
-    log.info("전시 정보 생성 성공 - id:{}, imageUrl:{}, status:{}", exhibition.getId(), imageUrl, status);
+    log.info(
+        "전시 정보 생성 성공 - id:{}, username:{}, status:{}",
+        exhibition.getId(),
+        exhibition.getUser().getUsername(),
+        status);
 
-    return exhibition.getId().toString() + "번 식별자의 전시가 성공적으로 생성되었습니다.";
+    return exhibitionMapper.toExhibitionDetailResponse(exhibition);
   }
 
   @Override
-  public PageResponse<ExhibitionResponse> getExhibitionPage(Pageable pageable) {
+  public ExhibitionDetailResponse getExhibition(Long id) {
+
+    Exhibition exhibition =
+        exhibitionRepository
+            .findById(id)
+            .orElseThrow(() -> new CustomException(ExhibitionErrorCode.EXHIBITION_NOT_FOUND));
+
+    return exhibitionMapper.toExhibitionDetailResponse(exhibition);
+  }
+
+  @Override
+  public Integer getExhibitionDraftCount() {
+
+    return exhibitionRepository
+        .findByUserIdAndFillAll(userService.getCurrentUser().getId(), false)
+        .size();
+  }
+
+  @Override
+  public PageResponse<ExhibitionResponse> getExhibitionPageByType(
+      SortBy sortBy, Pageable pageable) {
+    Page<ExhibitionResponse> page;
+
+    switch (sortBy) {
+      case HOTTEST:
+        page =
+            exhibitionRepository
+                .findAllOrderByLikesCountDesc(pageable)
+                .map(exhibitionMapper::toExhibitionResponse);
+        log.info("인기순 전시 리스트 페이지 조회 성공");
+        break;
+
+      case LATEST:
+        LocalDate cutoffDate = LocalDate.now().minusDays(7);
+        page =
+            exhibitionRepository
+                .findRecentOngoingExhibitions(cutoffDate, ExhibitionStatus.ONGOING, pageable)
+                .map(exhibitionMapper::toExhibitionResponse);
+        log.info("최신순 전시 리스트 페이지 조회 성공");
+        break;
+
+      default:
+        throw new CustomException(ExhibitionErrorCode.INVALID_SORT_TYPE);
+    }
+
+    return pageMapper.toExhibitionPageResponse(page);
+  }
+
+  @Override
+  public PageResponse<ExhibitionResponse> getExhibitionPageByUser(
+      Boolean fillAll, Pageable pageable) {
     User user = userService.getCurrentUser();
 
     Pageable sortedPageable =
@@ -89,10 +152,96 @@ public class ExhibitionServiceImpl implements ExhibitionService {
 
     Page<ExhibitionResponse> page =
         exhibitionRepository
-            .findByUserId(user.getId(), sortedPageable)
+            .findByUserIdAndFillAll(user.getId(), fillAll, sortedPageable)
             .map(exhibitionMapper::toExhibitionResponse);
 
-    log.info("{} 사용자의 전시 리스트 페이지 조회 - 호출된 페이지: {}", user.getNickname(), pageable.getPageNumber());
+    log.info(
+        "{} 사용자의 전시 리스트 페이지 조회 - 호출된 페이지: {}, 등록 완료 여부: {}",
+        user.getNickname(),
+        pageable.getPageNumber(),
+        fillAll);
     return pageMapper.toExhibitionPageResponse(page);
+  }
+
+  @Override
+  @Transactional
+  public ExhibitionDetailResponse updateExhibition(
+      Long id, MultipartFile image, ExhibitionRequest request) {
+
+    Exhibition exhibition =
+        exhibitionRepository
+            .findById(id)
+            .orElseThrow(() -> new CustomException(ExhibitionErrorCode.EXHIBITION_NOT_FOUND));
+
+    User currentUser = userService.getCurrentUser();
+    if (!exhibition.getUser().equals(currentUser)) {
+      throw new CustomException(ExhibitionErrorCode.EXHIBITION_ACCESS_DENIED);
+    }
+
+    if (request.getStartDate().isAfter(request.getEndDate())) {
+      throw new CustomException(ExhibitionErrorCode.INVALID_DATE_RANGE);
+    }
+
+    String imageUrl;
+
+    if (image != null) {
+      String newImageUrl = s3Service.uploadFile(PathName.EXHIBITION, image);
+
+      if (exhibition.getThumbnailImageUrl() != null) {
+        s3Service.deleteFile(s3Service.extractKeyNameFromUrl(exhibition.getThumbnailImageUrl()));
+      }
+      imageUrl = newImageUrl;
+    } else {
+      imageUrl = exhibition.getThumbnailImageUrl();
+    }
+
+    try {
+      ExhibitionStatus status = determineStatus(request.getStartDate(), request.getEndDate());
+
+      List<ExhibitionParticipant> participants = buildParticipants(request.getUserIdList());
+
+      participants.forEach(p -> p.setExhibition(exhibition));
+
+      Exhibition updatedExhibition =
+          exhibitionMapper.toExhibition(imageUrl, request, status, currentUser, participants);
+
+      exhibition.update(updatedExhibition);
+    } catch (Exception e) {
+      log.error("오류 로그: ", e);
+      throw new CustomException(ExhibitionErrorCode.EXHIBITION_API_ERROR);
+    }
+
+    log.info(
+        "전시 정보 수정 성공 - id: {}, status: {}", exhibition.getId(), exhibition.getExhibitionStatus());
+    return exhibitionMapper.toExhibitionDetailResponse(exhibition);
+  }
+
+  private ExhibitionStatus determineStatus(LocalDate startDate, LocalDate endDate) {
+    if (startDate == null || endDate == null) {
+      throw new CustomException(ExhibitionErrorCode.INVALID_DATE_RANGE);
+    }
+    LocalDate now = LocalDate.now(ZoneId.of("Asia/Seoul"));
+
+    if (endDate.isBefore(now) || endDate.isEqual(now)) {
+      return ExhibitionStatus.ENDED;
+    } else if (startDate.isBefore(now) || startDate.isEqual(now)) {
+      return ExhibitionStatus.ONGOING;
+    } else {
+      return ExhibitionStatus.UPCOMING;
+    }
+  }
+
+  private List<ExhibitionParticipant> buildParticipants(List<Long> userIds) {
+    return userIds.stream()
+        .distinct()
+        .map(
+            userId -> {
+              User participantUser =
+                  userRepository
+                      .findById(userId)
+                      .orElseThrow(() -> new CustomException(UserErrorCode.USER_NOT_FOUND));
+              return ExhibitionParticipant.builder().exhibition(null).user(participantUser).build();
+            })
+        .collect(Collectors.toList());
   }
 }
